Cleaned up data factories, readme, env variables
- Flattened factory folder structure: factories/posts/ → factories/ to keep things simpler, since we don't need subfolders yet
- Updated documentation 
- Removed name parameter, which was introduced as identifier of a particular factory, so that those factories can be filtered and deleted later. Since we don't use that model anymore, I have removed this
- Added custom number of workers per env variable, to custom parallelization locally
