+++ e2e/.env.example
@@ -1,2 +1,8 @@
 E2E_ACCOUNT_EMAIL=example
 E2E_ACCOUNT_PASSWORD=example
+
+# Debug failed tests (keeps containers)
+PRESERVE_ENV=false
+
+# define only if you want custom number of workers for tests to run in parallel locally
+TEST_WORKERS_COUNT=5
