+++ e2e/data-factory/README.md
@@ -6,21 +6,17 @@ A minimal test data factory for Ghost e2e tests, written in TypeScript.
 
 ```
 e2e/data-factory/          # Source files (TypeScript) - committed to git
-├── factories/
-│   ├── factory.ts         # Base factory class
-│   └── posts/
-│       └── post-factory.ts # Post factory implementation
+├── factory.ts             # Base factory class
+├── factories/             # Factory implementations
+│   ├── post-factory.ts
+│   ├── tag-factory.ts
+│   └── user-factory.ts
 ├── persistence/
 │   ├── adapter.ts         # Persistence interface
-│   └── adapters/
-│       └── knex.ts        # Knex SQL adapter
+│   └── adapters/          # Adapter implementations (API, Knex, etc)
+├── setup.ts               # Setup helper functions
 ├── index.ts               # Main exports
-├── utils.ts               # Utility functions
-└── tests/
-    └── post-factory.test.js # Test file (JavaScript)
-
-e2e/build/data-factory/    # Compiled files (JavaScript) - gitignored
-└── ... (generated by TypeScript compiler)
+└── utils.ts               # Utility functions
 ```
 
 ## Setup
@@ -47,73 +43,73 @@ This is part of the Ghost e2e test suite. All dependencies are managed by the ma
 
 ### In Tests
 
-```javascript
-// Import from the built files
-const {PostFactory, KnexPersistenceAdapter} = require('../build/data-factory');
+**Option 1: Use setup helpers (recommended)**
+```typescript
+import {createPostFactory, PostFactory} from '../data-factory';
 
-// Set up
-const db = knex(dbConfig);
-const adapter = new KnexPersistenceAdapter(db);
-const postFactory = new PostFactory(adapter);
+// Create factory with API persistence
+const postFactory: PostFactory = createPostFactory(page.request);
 
-// Build in memory (not persisted)
+// Build in-memory only (not persisted)
 const draftPost = postFactory.build({
     title: 'My Draft',
     status: 'draft'
 });
 
-// Create and persist
+// Create and persist to database
 const publishedPost = await postFactory.create({
     title: 'My Published Post',
     status: 'published'
 });
 ```
 
-### Running Tests
+**Option 2: Manual setup**
+```typescript
+import {PostFactory} from '../data-factory/factories/post-factory';
+import {GhostAdminApiAdapter} from '../data-factory/persistence/adapters/ghost-api';
 
-```bash
-# From e2e directory
-yarn test:factory
+const adapter = new GhostAdminApiAdapter(page.request, 'posts');
+const postFactory = new PostFactory(adapter);
 
-# This will:
-# 1. Build the TypeScript files
-# 2. Run Playwright tests in the data-factory directory
-
-# Or manually:
-yarn build
-yarn playwright test data-factory
+// Now you can build or create
+const post = await postFactory.create({
+    title: 'My Published Post',
+    status: 'published'
+});
 ```
 
 ## Development
 
-### Build Process
-
-- **TypeScript source files** (`*.ts`) are in `e2e/data-factory/`
-- **Compiled JavaScript** (`*.js`) goes to `e2e/build/data-factory/`
-- The `e2e/build/` directory is gitignored
-- Always rebuild after making changes: `cd e2e && yarn build`
-
 ### Adding New Factories
 
 1. Create a new factory class extending `Factory<TOptions, TResult>`
-2. Implement the `build()` method
-3. Set `name` and `entityType` properties
-4. Export from `index.ts`
+2. Implement the `build()` method (returns in-memory object)
+3. Set `entityType` property (used for persistence)
+4. Create a setup helper in `setup.ts` for convenient usage in tests
 
 Example:
 ```typescript
-export class UserFactory extends Factory<Partial<User>, User> {
-    name = 'user';
-    entityType = 'users';
-    
-    build(options: Partial<User> = {}): User {
-        // Generate test data
+import {Factory} from '../factory';
+
+export class MemberFactory extends Factory<Partial<Member>, Member> {
+    entityType = 'members';
+
+    build(options: Partial<Member> = {}): Member {
+        return {
+            id: generateId(),
+            email: options.email || faker.internet.email(),
+            name: options.name || faker.person.fullName(),
+            // ... more fields
+        };
     }
 }
 ```
 
-## Notes
-
-- The test file (`test/post-factory.test.js`) is JavaScript because it's a standalone script
-- All factory code is TypeScript and must be compiled before use
-- The compiled files in `e2e/build/` should never be committed to git
+Then create a setup helper:
+```typescript
+// In setup.ts
+export function createMemberFactory(httpClient: HttpClient): MemberFactory {
+    const adapter = new GhostAdminApiAdapter(httpClient, 'members');
+    return new MemberFactory(adapter);
+}
+```
