+++ e2e/data-factory/factories/post-factory.ts
@@ -1,7 +1,6 @@
 import {Factory} from '../factory';
-import type {PersistenceAdapter} from '../../persistence/adapter';
 import {faker} from '@faker-js/faker';
-import {generateId, generateUuid, generateSlug} from '../../utils';
+import {generateId, generateUuid, generateSlug} from '../utils';
 
 export interface Post {
     id: string;
@@ -34,12 +33,7 @@ export interface Post {
 }
 
 export class PostFactory extends Factory<Partial<Post>, Post> {
-    name = 'post'; // Factory identifier
-    entityType = 'posts'; // Entity name (for adapter; currently the database table)
-
-    constructor(adapter?: PersistenceAdapter) {
-        super(adapter);
-    }
+    entityType = 'posts'; // Entity name (for adapter; currently API endpoint)
 
     build(options: Partial<Post> = {}): Post {
         const now = new Date();
