+++ e2e/data-factory/factory.ts
@@ -1,7 +1,6 @@
-import type {PersistenceAdapter} from '../persistence/adapter';
+import type {PersistenceAdapter} from './persistence/adapter';
 
 export abstract class Factory<TOptions extends Record<string, unknown> = Record<string, unknown>, TResult = TOptions> {
-    abstract name: string;
     abstract entityType: string;
 
     protected adapter?: PersistenceAdapter;
