+++ e2e/data-factory/setup.ts
@@ -1,7 +1,7 @@
-import {PostFactory} from '../factories/posts/post-factory';
-import {TagFactory} from '../factories/tags/tag-factory';
-import {GhostAdminApiAdapter} from '../persistence/adapters/ghost-api';
-import {HttpClient} from '../persistence/adapters/http-client';
+import {PostFactory} from './factories/post-factory';
+import {TagFactory} from './factories/tag-factory';
+import {GhostAdminApiAdapter} from './persistence/adapters/ghost-api';
+import {HttpClient} from './persistence/adapters/http-client';
 
 /**
  * Create a new PostFactory with API persistence
