+++ e2e/helpers/pages/admin/index.ts
@@ -15,3 +15,4 @@ export * from './settings/SettingsPage';
 export * from './tags/TagsPage';
 export * from './tags/TagEditorPage';
 export * from './tags/TagDetailsPage';
+export * from './tags/NewTagsPage';
