+++ e2e/package.json
@@ -12,7 +12,6 @@
     "prepare": "tsc --noEmit",
     "pretest": "echo '⚠️ To make sure Ghost project is built with latest changes, run before tests: yarn docker:build (from repository root)'",
     "test": "playwright test --project=main",
-    "test:factory": "playwright test --project=factories",
     "test:single": "playwright test --project=main -g",
     "test:debug": "playwright test --project=main --headed --timeout=60000 -g",
     "test:types": "tsc --noEmit",
